feat: (BO) - exc/rtn 리팩토링
issue: #128
