+++ src/module/bo/exc/services/excUtils.js
@@ -1,26 +1,3 @@
-const excTypeMapping = {
-  1: '상체',
-  2: '하체',
-  3: '전신',
-  4: '유산소',
-}
-
-const mapExcType = type => {
-  return excTypeMapping[type] || '알 수 없음'
-}
-
-const excDifficultyMapping = {
-  1: '초급',
-  2: '초중급',
-  3: '중급',
-  4: '중상급',
-  5: '상급',
-}
-
-const mapDifficultyType = type => {
-  return excDifficultyMapping[type] || '알 수 없음'
-}
-
 const excAreaMapping = {
   1: '광배근',
   2: '기립근',
@@ -43,8 +20,5 @@ const mapExcAreaType = type => {
 
 export default {
   mapExcAreaType,
-  mapDifficultyType,
-  mapExcType,
   excAreaMapping,
-  excDifficultyMapping,
 }
